ui_test/api_test/requirements_offline.txt
attrs-23.1.0-py3-none-any.whl
certifi-2023.5.7-py3-none-any.whl
charset_normalizer-3.1.0-cp39-cp39-manylinux_2_17_aarch64.manylinux2014_aarch64.whl
fake_useragent-1.1.3-py3-none-any.whl
idna-3.4-py3-none-any.whl
importlib_resources-5.12.0-py3-none-any.whl
iniconfig-2.0.0-py3-none-any.whl
Logbook-1.5.3.tar.gz
more_itertools-9.1.0-py3-none-any.whl
numpy-1.25.0-cp39-cp39-manylinux_2_17_aarch64.manylinux2014_aarch64.whl
pytz-2023.3-py2.py3-none-any.whl
six-1.16.0-py2.py3-none-any.whl
python_dateutil-2.8.2-py2.py3-none-any.whl
packaging-23.1-py3-none-any.whl
pandas-1.3.5-cp39-cp39-manylinux_2_17_aarch64.manylinux2014_aarch64.whl
pluggy-0.13.1-py2.py3-none-any.whl
py-1.11.0-py2.py3-none-any.whl
pytest-6.0.2-py3-none-any.whl
pytest_html-2.1.1-py2.py3-none-any.whl
pytest_metadata-2.0.4-py3-none-any.whl
pytest_rerunfailures-11.1.2-py3-none-any.whl
PyYAML-6.0-cp39-cp39-manylinux_2_17_aarch64.manylinux2014_aarch64.whl
requests-2.30.0-py3-none-any.whl
toml-0.10.2-py2.py3-none-any.whl
urllib3-2.0.3-py3-none-any.whl
xlrd-1.2.0-py2.py3-none-any.whl
xlutils-2.0.0-py2.py3-none-any.whl
xlwt-1.3.0-py2.py3-none-any.whl
zipp-3.15.0-py3-none-any.whl
